id_codec, categories 코드 작성 및 common/reports파일 생성 및 작성, init파일 수정
